+++ src/ResourceManager/Insights/Commands.Insights/Properties/AssemblyInfo.cs
@@ -24,5 +24,5 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 [assembly: CLSCompliant(false)]
 [assembly: Guid("ed102280-3577-49bf-93dd-11b6e3a44a57")]
-[assembly: AssemblyVersion("5.1.5")]
-[assembly: AssemblyFileVersion("5.1.5")]
+[assembly: AssemblyVersion("5.1.6")]
+[assembly: AssemblyFileVersion("5.1.6")]