+++ src/ResourceManager/LogicApp/Commands.LogicApp/Properties/AssemblyInfo.cs
@@ -29,5 +29,5 @@ using System.Runtime.InteropServices;
 [assembly: CLSCompliant(false)]
 
 [assembly: Guid("e50b915e-1c94-4cd1-bbea-72ff2246d237")]
-[assembly: AssemblyVersion("4.1.4")]
-[assembly: AssemblyFileVersion("4.1.4")]
+[assembly: AssemblyVersion("4.1.5")]
+[assembly: AssemblyFileVersion("4.1.5")]